refactor(core): 重命名多个文件及其引用路径以统一命名规范
- 将核心模块文件名改为短横线风格（kebab-case）
- 更新所有相关导入路径以匹配重命名后的文件名
- 修改测试文件名和对应引用路径，保持一致性
- 调整部分组件及钩子中的类型导入路径
- 无业务功能变更，仅代码结构及命名优化
