+++ src/cli.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "ink";
 import { setShellIfWindows } from "./common/system/shell-utils";
-import { checkForNpmUpdate, promptForPendingUpdate, type PackageInfo } from "./common/updateCheck";
+import { checkForNpmUpdate, promptForPendingUpdate, type PackageInfo } from "./common/update-check";
 import { AppContainer } from "./ui";
 
 const args = process.argv.slice(2);
