+++ src/tests/prompt-input-keys.test.ts
