+++ src/tests/prompt-undo-redo.test.ts
@@ -8,7 +8,7 @@ import {
   recordPromptEdit,
   redoPromptEdit,
   undoPromptEdit,
-} from "../ui/core/promptUndoRedo";
+} from "../ui/core/prompt-undo-redo";
 
 test("prompt undo and redo restore edited buffer states", () => {
   const history = createPromptUndoRedoState();
