+++ src/tests/thinking-state.test.ts
