+++ src/tests/welcome-screen.test.ts
