+++ src/ui/core/loading-text.ts
