+++ src/ui/core/slash-commands.ts
