+++ src/ui/core/thinking-state.ts
