+++ src/ui/hooks/useHistoryNavigation.ts
@@ -1,6 +1,6 @@
 import type React from "react";
 import { useCallback, useState } from "react";
-import type { PromptBufferState } from "../core/promptBuffer";
+import type { PromptBufferState } from "../core/prompt-buffer";
 
 export type HistoryNavigationState = {
   historyCursor: number;
