+++ src/ui/views/AskUserQuestionPrompt.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, useState } from "react";
 import { Box, Text } from "ink";
-import type { AskUserQuestionAnswers, AskUserQuestionItem } from "../core/askUserQuestion";
+import type { AskUserQuestionAnswers, AskUserQuestionItem } from "../core/ask-user-question";
 import { useTerminalInput } from "../hooks";
 
 type Props = {
